requirements.txt
joblib==1.5.1
jpype1==1.5.2
mpmath==1.3.0
numdifftools==0.9.41
numpy==2.2.6
packaging==25.0
pandas==2.2.3
patsy==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
semopy==2.3.11
six==1.17.0
statsmodels==0.14.4
sympy==1.14.0
threadpoolctl==3.6.0
tzdata==2025.2
